logo code improved
